requirements.txt
google-genai
tinydb
python-dotenv
Pillow
gradio_client